refeat: updating some components and pages
