+++ src/app/globals.css
@@ -9,4 +9,20 @@
         padding: theme('spacing.10') theme('spacing.20');
         box-shadow: theme('boxShadow.md');
       }
+      .form{
+        margin-top: 2rem;
+        width: 100%;
+        gap: 1.25rem;
+        display: flex;
+        flex-wrap: wrap;
+        justify-content: center;
+        padding: theme('spacing.10') theme('spacing.20');
+        border-radius: theme('borderRadius.2xl');
+      }
+      .container {
+        margin-bottom: 1.5rem;
+        margin-top: 13rem;
+        display: flex;
+        justify-content: center;
+      }
 }
