+++ src/components/COMPONENTES/NormalInput.tsx
@@ -1,42 +1,69 @@
-import React from 'react';
+import React, { useState } from 'react';
 
-type NormalInputType = {
-    label: string,
-    id: string,
-    placeholder: string
-    disabled?: boolean,
-	value?: string,
-	onChange?(e:any):void
+import { FiEye, FiEyeOff } from 'react-icons/fi';
+
+interface CustomInputProps extends React.InputHTMLAttributes<HTMLInputElement> {
+	label: string;
 }
 
-const NormalInput = ({
-    id,
-    label,
-    placeholder,
-    disabled,
-	onChange,
-	value,
-}: NormalInputType) => {
+const NormalInput: React.FC<CustomInputProps> = ({ label, ...inputProps }) => {
+	const [passwordVisible, setPasswordVisible] = useState(false);
+	const [password, setPassword] = useState('');
+	const handleTogglePasswordVisibility = () => {
+		setPasswordVisible(!passwordVisible);
+	};
 	return (
-		<div className="mb-5 flex w-[45%] flex-col">
-			<label className="mb-2 text-sm font-medium" htmlFor={id}>
-				{label}
-			</label>
+		!inputProps.hidden &&
+		(inputProps.type == 'password' ? (
+			<>
+				<div className="mb-5 flex w-[45%] flex-col">
+					<label className="mb-2 text-sm font-medium" htmlFor="password">
+						{label}
+					</label>
+					<div className="flex items-center justify-between rounded-md border border-gray-300 bg-white px-4 py-2.5">
+						<input
+							{...inputProps}
+							className="w-11/12 rounded-md border-0 bg-white text-sm outline-none"
+							type={passwordVisible ? 'text' : 'password'}
+							value={password}
+							onChange={(e) => setPassword(e.target.value)}
+							name={inputProps.id}
+						/>
+						{passwordVisible ? (
+							<FiEye
+								className="h-4 w-4 text-black"
+								onClick={handleTogglePasswordVisibility}
+							/>
+						) : (
+							<FiEyeOff
+								className="h-4 w-4 text-black"
+								onClick={handleTogglePasswordVisibility}
+							/>
+						)}
+					</div>
+				</div>
+			</>
+		) : (
+			<>
+				<div className="mb-5 flex w-[45%] flex-col">
+					<label className="mb-2 text-sm font-medium" htmlFor={inputProps.id}>
+						{label}
+					</label>
 
-			<div className="rounded-md border border-gray-300 bg-white px-4 py-2">
-				<input
-					className="w-full rounded-md border-0 bg-white text-sm outline-none"
-					type='text'
-					name={id}
-					id={id}
-					placeholder={placeholder}
-					required
-					disabled={disabled}
-					onChange={onChange}
-					value={value}
-				/>
-			</div>
-		</div>
+					<div
+						className={`w-full rounded-md border border-gray-300 bg-white px-4 py-2`}
+					>
+						<input
+							className={`w-full rounded-md border-0 bg-white text-sm outline-none `}
+							{...inputProps}
+							name={inputProps.id}
+							onChange={inputProps.onChange}
+							value={inputProps.value}
+						/>
+					</div>
+				</div>
+			</>
+		))
 	);
 };
