+++ src/components/COMPONENTES/Title.tsx
@@ -3,16 +3,22 @@ import React from 'react';
 type TitleType = {
     title: string,
     subtitle: string,       
-    color: string,       
+    colorHex: string,       
 }
 const Title = ({
     title,
     subtitle,
-    color
+    colorHex
 }: TitleType) => {
+
 	return (
 		<div className='mb-10'>
-			<h1 className={`text-center text-3xl font-bold text-[${color}]`}>
+			<h1 
+			style={{color: colorHex}}
+			className={`
+			text-center text-3xl font-bold 
+			text-[${!colorHex && ('black')}]
+			`}>
 				{title}
 			</h1>
